+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to `NovaPoshta` will be documented in this file
 
+### 2020-10-22
+- [Fix] `getWarehouses` in Address ($cityRef = null)
+
+
 ### 2020-10-27
 - [Add][ND] `getCatalogCounterparty` in Counterparty
 - [Add] `getSettlements` in Address